doc/test; Updates docs and test MeshModel
test: MeshModel and Mesh
